Further address comments in PR re: incorporation of RefSeq IDs into SequenceDictioanry
