+++ adam-cli/src/main/scala/org/bdgenomics/adam/cli/Vcf2ADAM.scala
@@ -23,9 +23,7 @@ import org.apache.hadoop.mapreduce.Job
 import org.apache.spark.{ Logging, SparkContext }
 import org.apache.spark.rdd.RDD
 import org.kohsuke.args4j.{ Option => Args4jOption, Argument }
-import java.io.{ FileOutputStream, File }
-import net.sf.samtools.SAMFileReader
-import org.apache.commons.io.IOUtils
+import java.io.File
 
 object Vcf2ADAM extends ADAMCommandCompanion {
   val commandName = "vcf2adam"
@@ -47,32 +45,12 @@ class Vcf2ADAMArgs extends Args4jBase with ParquetArgs with SparkArgs {
   var outputPath: String = null
 }
 
-class Vcf2ADAM(val args: Vcf2ADAMArgs) extends ADAMSparkCommand[Vcf2ADAMArgs] with Logging {
+class Vcf2ADAM(val args: Vcf2ADAMArgs) extends ADAMSparkCommand[Vcf2ADAMArgs] with DictionaryCommand with Logging {
   val companion = Vcf2ADAM
 
   def run(sc: SparkContext, job: Job) {
 
-    def getDictionaryFile(name: String): Option[File] = {
-      val stream = ClassLoader.getSystemClassLoader.getResourceAsStream("dictionaries/" + name)
-      if (stream == null)
-        return None
-      val file = File.createTempFile(name, ".dict")
-      file.deleteOnExit()
-      IOUtils.copy(stream, new FileOutputStream(file))
-      Some(file)
-    }
-
-    def getDictionary(file: File) = Some(SequenceDictionary(SAMFileReader.getSequenceDictionary(file)))
-
-    var dictionary: Option[SequenceDictionary] = None
-    if (args.dictionaryFile != null) {
-      if (args.dictionaryFile.exists) {
-        dictionary = getDictionary(args.dictionaryFile)
-      } else getDictionaryFile(args.dictionaryFile.getName) match {
-        case Some(file) => dictionary = getDictionary(file)
-        case _          => assert(false, "Supplied dictionary path is invalid")
-      }
-    }
+    var dictionary: Option[SequenceDictionary] = loadSequenceDictionary(args.dictionaryFile)
     if (dictionary.isDefined)
       log.info("Using contig translation")
